second stage update
