+++ Módulo 1/Avaliações/Primeira Avaliação/Segunda Etapa/Common.h
@@ -1,0 +1,16 @@
+#ifndef COMMON_H
+#define COMMON_H
+
+#include <string>
+
+struct Date
+{
+    int day, month, year;
+};
+
+struct Hour
+{
+    int hour, minute;
+};
+
+#endif
